Add initial working agreement proposal
This is a proposal for a working agreement, mostly focused on how we share and review code using GitHub and pull requests.  The intent is to capture the most important aspects of how we work as a team, and have a process for discussing this and revising it so that everyone on the team has a shared understanding about how we work well together, and why we work the way we do.